Split sdo_client in two files : main and helpers Added reading Uint8 through Uint64
